refactor: 70 - change placement of select component
